api/requirements.txt
asgiref==3.5.2
attrs==22.1.0
coverage==6.5.0
Django==4.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==4.0.0
pytest-django==4.5.2
pytz==2022.5
sqlparse==0.4.3
tomli==2.0.1
